Initial predexp impl

+++ src/main/util/conversions.cc
@@ -100,7 +100,7 @@ int get_int_property(int* intp, Local<Object> obj, char const* prop, const LogIn
 		as_v8_error(log, "Type error: %s property should be integer", prop);
 		return AS_NODE_PARAM_ERR;
 	}
-	(*intp) = value->IntegerValue();
+	(*intp) = (int) value->IntegerValue();
 	as_v8_detail(log, "%s => (int) %d", prop, *intp);
 	return AS_NODE_PARAM_OK;
 }
@@ -111,7 +111,7 @@ int get_optional_int_property(int* intp, bool* defined, Local<Object> obj, char 
 	Local<Value> value = obj->Get(Nan::New(prop).ToLocalChecked());
 	if (value->IsNumber()) {
 		if (defined != NULL) (*defined) = true;
-		(*intp) = value->IntegerValue();
+		(*intp) = (int) value->IntegerValue();
 		as_v8_detail(log, "%s => (int) %d", prop, *intp);
 	} else if (value->IsUndefined() || value->IsNull()) {
 		if (defined != NULL) (*defined) = false;
